+++ etl-workers/activities-worker/src/main/java/com/example/workers/activities/ExecutionActivities.java
@@ -4,8 +4,8 @@ import io.temporal.activity.ActivityInterface;
 
 @ActivityInterface
 public interface ExecutionActivities {
-    String start(int id);
-    String end(int id);
-    String terminate(int id);
+    String start(long id);
+    String end(long id);
+    String terminate(long id);
     String ping();
 }
